update-fit-driver
fit_vae will now read the configuration file to determine whether to use a traditional vae or the convolutional vae. Additionally, it will check to see if there is already a model saved at the name in the config file and if so it will load that model and then save additional training to a new file with '_copy' as a suffix in the filename. Also no longer wastefully loads the HDF data file just to check the dimensions, since those are in the config file as well